backend/requirements.txt
fastapi
uvicorn
httpx
pycryptodome
python-multipart
